pipeio: Add fwts_pipe_open_rw
This change introduces a function to open a process pipe that supports
both reading and writing (ie, we open pipes for both stdin and stdout of
the child process).

We do this by extending and renaming the existing fwts_pipe_open_ro,
then use the new function to implement fwts_pipe_open_ro.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>