bios: mtrr: fix cache type corner case at start of mtrr range
To assign a cache type to a memory range, we determine if _any_ part of
it is covered by one or more of the mtrr list entries.

If only the last location of the memory range under test lies inside
the range described by the mttr entry, we will fail to identify it. As an
example, consider that the mttr range is from [0x2000, 0x3000) - inclusive
of 0x1000 but exclusive of 0x2000. If the range under test is [0x1000,
0x2000], the second test of the condition will not be met. Thus, it is
necessary to check whether the end of the memory range under test is less
_or_ equal to the start of the range given by the mtrr entry.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Acked-by: Colin Ian Jing <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>