V3 Introduce IPMI BMC Info
This feature adds the foundation to perform an IPMI BMC Info
check that will determine if a Host is capable of
IPMI messaging and if so will perform a basic IPMI message
exchange to determine the version of IPMI running
on the hardware.  In the future the IPMI infrastructure
can be used to further interrogate the FRU Inventory and other
Sensors to help correlate data and surface any
discrepancies on inventory or hardware characteristics.

Signed-off-by: Deb McLemore <debmc@linux.vnet.ibm.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>