devicetree/dt_sysinfo: Add OPAL reference compatible checks
We added a new device tree check for the OPAL compatible property.
We also added an OpenPOWER reference check to validate that the
machine model and device tree compatible property are a supported
match.  Future additions will add device tree hardware checks
(as applicable) to confirm that the device tree matches the topology of
the expected hardware configurations.

Signed-off-by: Deb McLemore <debmc@linux.vnet.ibm.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>