devicetree/dt_sysinfo: clean up logic in stringlist_contains_last
This change is a little cleanup of dt_fdt_stringlist_contains_last, but
retains the same functional semantics.

We remove the duplicated calls to memcmp, instead only changing the
arguments if !p. Also, we return a bool instead of an int, and change
the prefix from prd_ to dt_, as this is a device-tree helper (and not a
prd one).

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>