devicetree/dt_sysinfo: fix reference platform compatible vs. model check
The logic in machine_matches_reference_model is slightly incorrect; if
we're *not* running on a reference platform, then we don't want to fail.

The check should only fail if the compatible string reports a reference
platform, but the model number does not.

This change reworks the logic in machine_matches_reference model to
reflect this, allowing us to simplify it a little.

Reported-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>