fwts_cpuinfo_x86: check for null pointers when using
When successful, fwts_cpu_get_info returns a non-null pointer to a structure
containing the CPU information. However, this does not imply that all the
members of the structure are valid. This can happen if, for instance,  the
requested CPU or one of its properties was not found.

Thus, the sane thing to do is to have the callers of fwts_cpu_get_info to
make sure that data is valid before using it. Otherwise, segmentation faults
can occur if callers try to dereference NULL pointers from the CPU info
structure.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>