lib: fwts_cpu: query info from the first online CPU
All calls to fwts_cpu_get_info request the info of CPU0. However, this CPU
may be offline if the Linux kernel was built with
CONFIG_BOOTPARAM_HOTPLUG_CPU0. In such a case, fwts_cpu_get_info would
return an empty (and useless) structure.

Assuming that all the CPUs yield the same information, we could use the
first online CPU to query its information. We already operate under this
assumption by always querying the CPU0 info.

Functionality to query a specific CPU is preserved but now it is possible
to use -1 to query the first available CPU.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>