efi: enable module loading to load legacy or new efi driver
The name of the efi module and the device is going to change once
the driver lands upstream in the kernel.  So make the module loading
more flexible to cater for legacy and the new upstream'd kernel driver.

Also abstract out the efi test device driver name and provide a thin
open/close device abstraction layer so we can keep the device name
out of the tests.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>