fwts/madt: Implement I/O APIC and I/O SAPIC value comparison test
If both I/O APIC and I/O SAPIC subtables exist, there are two checks that
should be run.  The first is to confirm that there are at least as many
I/O SAPIC tables as there are I/O APIC tables.  The second is to check that
every I/O APIC ID has a corresponding I/O SAPIC ID table.

This patchset implements the tests.

Additionally note a fixed typo for "MADIOSPAICAddrZero".

[v2]: Fix off-by-one error pointed out during review

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: colin.king@canonical.com
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>