acpi: madt: Use 20 as the GIC ITS node length
The ACPI 6.0 spec technically defines the GIC ITS node length as 16,
however the conbination of all defined fields is 20 bytes, so the length is
really 20.  The spec defined length is corrected in following revisions of
the spec.

20 is what ACPI implementations are expected to be using in practice, so
updating the test to use 20 despite the spec definition has the test match
the real world, updates the test implementation to match the test
documentation, and allows the test to correctly find the start of any nodes
following a GIC ITS node in the MADT table.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>