fwts/madt: Add processor UID checking to madt tests
The current madt test does not do any processor UID comparisons.  They are
marked as TODOs in the file.

This patchset implements two ACPI searches (one for old-style Processor
Declaration operator, and another for the new Processor Device operator)
that populate a list of UID entries, which can be queried for each MADT
struct's UID.

[v2]: Correctly evaluate MADT LSAPIC table ACPI processor ID,  UID
value, and UID string.
[v3]: Fix typo in MADTLSAPICUidCompare message.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>