acpi/nfit: Fix endless loop on broken NFIT tables
When running 'fwts nfit' on a system with an incorrect subtable length of
zero, the nfit test will loop endlessly.

This results.log contains many entries of

NFIT NVDIMM Firmware Interface Table:
  Reserved:                 0x00000000

  NFIT Subtable:
    Type:                                   0x0000
    Length:                                 0x0000
    SPA Range Structure Index:              0x0000
    Flags:                                  0x0000
    Reserved:                               0x00000000
    Proximity Domain:                       0x00000000
    Address Range Type GUID:                00000000-0000-0000-0000-000000000000
    System Physical Address Range Base:     0x0000000000000000
    System Physical Address Range Length:   0x0000000000000000
    Address Range Memory Mapping Attribute: 0x0000000000000000
FAILED [HIGH] NFITBadRangeIndexZero: Test 1, NFIT SPA Range Structure Index must
not be zero

This occurs because the test assumes a valid table length.  While the ACPI
specification is not explicit in indicating that a zero length is invalid,
it certainly is implied that it cannot be zero.

This patch adds a check and aborts the NFIT test on a zero subtable length.

As a result the output of the test is now

NFIT NVDIMM Firmware Interface Table:
  Reserved:                 0x00000000

  NFIT Subtable:
    Type:                                   0x0000
    Length:                                 0x0000
FAILED [HIGH] NFITLengthZero: Test 1, NFIT Subtable (offset 0x28) length cannot
be 0

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: Alex Hung <alex.hung@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>