acpi: bert: do not fail if BERT error region raw_data_length is zero
If there are no errors to report, the Boot Error Region pointed by the
BERT ACPI table in certain systems is populated with zeros. This implies
that the Raw Data Offset field will also be zero. In such a case, we will
not meet the condition

    error_region.raw_data_offset > sizeof(error_region) +
                                   error_region.data_length).           (1)

This is not an error if there is not raw data. Instead, fail the test case
only if error_region.raw_data_length is non-zero.

This brings FWTS in alignment with what Linux does to check the validity of
the BERT error region.

In order to not mask actual issues in the BERT error region and conserve
some of the current behavior, issue a warning when (1) is not met and add
commentary on the fact that there might not be raw data.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>