acpi: fadt: Remove HEADLESS check on reduced hardware
The ACPI Spec technically states that the HEADLESS and reduced hardware
flags are mututally exclusive, however the Spec provides no indication why
a reduced hardware platform could not be HEADLESS.  There currently are
OSes such as VMware's ESXi which expect HEADLESS to be set where
appropiate, regardless of the reduced hardware flag.

Lets not penalize platforms for doing what is expected in the real world.
We expect a future revision of the ACPI Spec to allow HEADLESS to be set
with reduced hardware due to the existance of this real world usecase.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>