uefirttime: specify enable to true for setwakeuptime wiht null time test
UEFI spec. allows the Time parameter for SetWakeupTime to be NULL when the
enable is false, so here should set enable to true for testing NULL time
parameter.

>From spec.
typedef EFI_STATUS
SetWakeupTime (IN BOOLEAN Enable, IN EFI_TIME *Time OPTIONAL);

If Enable is TRUE, the time to set the wakeup alarm for. Type
EFI_TIME is defined in the GetTime() function description.
If Enable is FALSE, then this parameter is optional, and may be
NULL.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>