pci: aspm: Add segment support
PCI segement groups are the first level of relationship between PCI devices
in a system.  PCI segments allow bus numbers to be reused as a particular
bus ID must be unique only within the segment upon which it resides.

Segment support is important as a system may have multiple root ports and
attached devices across multiple segements, with different ASPM settings
for each port/device set.  If the test is not aware of segments, it may
incorrectly match a root port from segment A with a device from segment B
and cause a test failure if that port and device happen to not have
matching ASPM settings.

Change-Id: I2aea72d12ca1dc945a4f73cf5cc652c6ffc62654
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>