acpi: madt: Fix processor UID check
Commit 763737164714 ("fwts/madt: Add processor UID checking to madt tests")
is broken on the Qualcomm Technologies QDF2432 and likely very fragile on
other platforms as well.

The root cause is that the added redefinition of the acpi_integer structure
is not compatible with AcpiEvaluateObject() and can result in
AE_BUFFER_OVERFLOW errors.

AcpiEvaluateObject() ensures that the provided return buffer is large enough
to hold the resulting object.  The size of the resulting object is
determined via AcpiUtGetSimpleObjectSize() which uses ACPI_OBJECT as a
baseline for the size of a resulting object (variable length objects like
strings may be larger).  Per src/acpica/source/include/actypes.h ACPI_OBJECT
is a union, and therefore the size of ACPI_OBJECT is atleast the size of its
largest member.  The Integer member is not the largest member, thus an
ACPI_OBJECT cannot fit into acpi_integer struct as defined in the madt test.
On QDF2432, sizeof(acpi_integer) is 16, where as sizeof(ACPI_OBJECT) is 24.
Since 24 > 16, AcpiEvaluateObject returns AE_BUFFER_OVERFLOW.

This results in false test failures because the test has not properly parsed
the processor device UIDs, and cannot later match the GICC UIDs to any
processor UIDs.

We fix this by not reinventing the wheel, and using ACPI_OBJECT as our
output buffer for AcpiEvaluateObject (both Processor and Integer objects
that we care about are fixed size), and parsing the union nativly for the
values we care about.  Since the source object is also ACPI_OBJECT, the
output buffer will always be of the same type, and thus the same size,
preventing AE_BUFFER_OVERFLOW errors, particularly if ACPI_OBJECT grows in
size in the future (ie an added member, or some stange compiler packing).

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>