lib: olog: Show test as skipped if nothing to do
If there is nothing for olog to do, the error message states the test is
skipped, however the test is categorized as aborted.  The majority of other
tests are categorized as skipped under this scenario.  Update the init path
to return a SKIP error to categorize the test result similar to other tests.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>