acpi: pmtt: Show test as skipped if table does not exist
If the table does not exist, the error message states the test is being
skipped, however the test is categorized as aborted.  The majority of other
tests are categorized as skipped under this scenario.  Update the init path
to return a SKIP error to categorize the test result similar to other tests.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>