acpi: dbg2: fix segfault and refactor dbg2_obj_find
The dbg2 test may crash on some systems due to dbg2_obj_find not being
able to handle malformed or nonstandard NamespaceString attributes.
It has been seen to crash due to NamespaceString being described with a
leading escape backslash, such as '\\_SB_.AHBC.URT1'. In these cases,
regardless of whether the test should pass or not, it should not crash
fwts.
In order to avoid more problems with exact length calculation, the
'expanded' string size is just moved to a reasonably large size which
should be able to fit any NamespaceString.

It has also been observed that the stao table test implements a very
similar routine and therefore may be vulnerable to the same sort of
problem, so that code has been refactored into fwts_acpi_obj_find which
can be used by both stao and dbg2 tests.

The printed strings in case of error logs are now printed with the
unexpanded string, so the expected outputs used in 'make check' have
also been updated to reflect that.

Signed-off-by: Erico Nunes <ernunes@redhat.com>
Acked-by: Alex Hung <alex.hung@canoniocal.com>
Acked-by: Colin Ian King <colin.king@canonical.com>