fwts_framework.h: fix redefinition typedef warning
When building the current fwts release with an older gcc 4.4, we get
this redefinition typedef warning on struct fwts_framework, due to a
forward typedef declaration which is made in the beginning of the file:

  In file included from ../../../src/lib/include/fwts_binpaths.h:27,
                   from ../../../src/lib/include/fwts.h:54,
                   from fwts_ac_adapter.c:19:
  ../../../src/lib/include/fwts_framework.h:158: error: redefinition of
  typedef ‘fwts_framework’
  ../../../src/lib/include/fwts_framework.h:28: note: previous
  declaration of ‘fwts_framework’ was here

By removing the second typedef in the definition, we can get rid of the
warning and still build with every gcc version.
As there are still supported systems with older gcc versions, such as
RHEL6, it is interesting to have this fixed.

This has been built and tested with gcc 6.2.1, gcc 4.8.5 and gcc 4.4.7.

Signed-off-by: Erico Nunes <ernunes@redhat.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>