fwts_hwinfo.c: fix strict-aliasing error with old gcc
When building the current fwts release with an older gcc 4.4, we get the
following error due to breaking strict-aliasing rules, despite the
comment in the code which states there has been an attempt to solve this
in the past:

  fwts_hwinfo.c: In function ‘fwts_hwinfo_get’:
  fwts_hwinfo.c:420: error: dereferencing pointer ‘sockaddr’ does break
  strict-aliasing rules
  fwts_hwinfo.c:419: note: initialized from here

This new attempt to fix it just makes a copy of the whole "buf.ifr_addr"
to into sockaddr to avoid running into any strict-aliasing problems.
As there are still supported systems with older gcc versions, such as
RHEL6, it is interesting to have this fixed.

This has been built and tested with gcc 6.2.1, gcc 4.8.5 and gcc 4.4.7.

Signed-off-by: Erico Nunes <ernunes@redhat.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>