esrt.c: fix uninitialized use of 'fwversion' warning
This warning shows up with different gcc versions, and may become a
problem due to the use of -Werror:

  uefi/esrt/esrt.c: In function ‘check_entries’:
  uefi/esrt/esrt.c:84: error: ‘fwversion’ may be used uninitialized in this
  function

The warning seems to point to what potentially can be a bug. So, just
initialize fwversion to a safe value to fix it.

This has been built and tested with gcc 6.2.1, gcc 4.8.5 and gcc 4.4.7.

Signed-off-by: Erico Nunes <ernunes@redhat.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>