fwts_framework: handle -? option differently from -h
When optargs fail, the optarg short option '?' is returned
which currently acts the same as -h.  The result is that
the user can use an incorrect option and they get more than
a console's worth of help information and the error message
scrolls off the screen, which isn't too helpful.  Instead,
make the -? option just inform the user that there is a -h
option for more help, thus the error message from the failed
optarg scanning is not scrolled off the screen.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>