fwts-alloc: garbage collect hash records
The hash table that tracks low memory allocations does not
automatically free records when they are not used (instead
it marks them as re-usable which is faster).  This patch
adds garbage collection to the hash table when the number
of free'd items drops down to zero. This lazy deferral
allows us to efficiently reap the hash when it is no longer
in use.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>