Integrate PPC for FWTS-LIVE Frontend
This feature adds the integration of the PPC platform for running
the FWTS-LIVE tests.

PPC available tests are queried and dynamic execution occurs on
a platform basis, i.e. only PPC tests are run when PPC architecture
is detected.  For BIOS and ACPI tests, these tests are the default
tests and are conditionally queried and executed based on the
uname of the runtime system.

Signed-off-by: Deb McLemore <debmc@linux.vnet.ibm.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>