lib: fwts_alloc: move all memory tracking to hash records
Previous changes to the low memory allocator allows us to track
each allocation with a hash table.  We now extend this to also
track the total allocation size; this allows us to remove the
allocation size tracking using a header at the start of each
allocation with a magic to sanity check it.  Separation of the
metadata from the data reduces the risk of a bad write to the
allocated low memory corrupting the metadata.

This change also means that low-memory allocations are now
page aligned (and hence cache-aligned) which is more optimal
in cache performance.

I also added some sanity checking to size checking for allocations,
and overflow checking.

Finally, I added malloc()/calloc()/realloc() compatible ENOMEM
setting to errno when we run out of memory (or when the allocator
detects something is broken).

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>