lib: fwts_alloc: fix incorrect accounting on hash entry count
When hash entries are re-used we also need to bump the count of active
entries to ensure we are tallying these correctly.  This fixes a
non-fatal issue where the accounting believes that all entries are freed
when there could be some still being active and we end up prematurely
cleaning up the hash table too early.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>