acpica: add some extra run time verification to FACP (FADT)
While running fwts through some regression tests from random broken
ACPI data dumps I found that some files have an FACP (FADT) that
point to the FACS or DSDT which actually don't exist in the data
dump. This causes ACPICA to segfault because we have pointers to
non-existent tables. Add a level of verification on these tables
to ensure we are not passing garbage to the ACPICA execution engine.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>