lib: fwts_acpi_tables: Add extra fix-up for FACS and DSDT table addresses
In the case where the FACS and DSDT are loaded from file and don't have
an address then fwts makes up a fake address for these tables to keep
ACPICA happy.  However, the FACP (FADT) points to these tables so we
can actually fix up the FACS and DSDT table addresses to these known
addresses fiven by the FACP.  We prefer 64 bit addresses from the FACP
but if they don't exist, use the 32 bit addresses.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>