lib: fwts_log: add LOG_NO_FIELD and use this to indicate an unfound field
Rather than implicitly using a non-enum fwts_log_field value zero
to indicate an unfound field instead use a new explicitly defined
LOG_NO_FIELD enum value to indicate this.  This cleans up a couple
of static analysis warnings of the form:

"The variable 'field' is of enum type. It is odd that it is used as a
variable of a Boolean-type."

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>