fwts-test: disable some acpi specific x86 tests for non-x86 systems
Due to the way ACPI tables can be faked when loading them in from file
we're getting some memory address differences on x86 compared to other
architectures. For the moment, let's disable these tests and I'll try
and figure out a better way to test these that don't need to concern
the different addressing layouts.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>