opal: fix tests dependency on libfdt
The inclusion of some tests in src/opal/ in the latest fwts release
added a hard dependency on libfdt-devel for all architectures, otherwise
breaking the build.
libfdt-devel may not be available or wanted for all architectures, for
example it is not officially distributed with RHEL x86_64.

fwts already had an optional dependency on libfdt before and this is
handled through the use of HAVE_LIBFDT.
This handling was already done for some of the opal tests but the new
ones were trying to build regardless of HAVE_LIBFDT. This patch moves
the new ones along with the tests which will only be built if
HAVE_LIBFDT gets set during configuration, as they rely heavily on it.

Signed-off-by: Erico Nunes <ernunes@redhat.com>
Cc: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>
Signed-off-by: Erico Nunes <ernunes@redhat.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>