lib: fwts_alloc: re-work the memory parsing again
Unfortunately the previous fix of mine causes tainted variable warnings
on CoverityScan static analysis.  I'm putting the original code back and
adding a file position check to fix the bug that the previous patch fixed.
The issue is that fscanf on the last line of a /proc/self/maps may cause
an infinite loop if it does not parse with the expected input and we don't
hit the EOF.  So instead, check that the parsing isn't stuck on file position
and if it has not moved forward then we break out of the loop.

Fixes: 56ca0bef8cea79 ("lib: fwts_alloc: only parse mem info once we have a valid line read in")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>