lib: fwts_safe_mem: remove need to copy into a buffer
While fwts_safe_memread() works fine as it is, it is copying data
to the stack and we don't guard how big that copy can be, so we
potentially could get a segfault if we run out of stack. Instead
just read the data. Force gcc not to optimize out the reads by
using volatile.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>