fwts_safe_mem: workaround longjmp clobber warnings
On some flavours of GCC the src to fwts_safe_memread throws up
a longjmp clobber warning:

fwts_safe_mem.c: In function ‘fwts_safe_memread’:
fwts_safe_mem.c:69:35: error: argument ‘src’ might be clobbered by ‘longjmp’
or ‘vfork’ [-Werror=clobbered]
 int fwts_safe_memread(const void *src, const size_t n)
                                   ^
cc1: all warnings being treated as errors

We either disable -Wclobber (which I don't want to do since it is useful
to catch these potential errors) or work out a case by case workaround.
The easiest fix for these is to crank down the optimisation level to zero
so that the compiler stashes the src argument onto the stack and we don't
end up with clobbering.

So.. add a bunch of macros to use a gcc attribute if it is supported
and force the function to be optimized at level 0.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>