lib: fwts_smbios: use fwts_safe_memcpy to read SMBIOS entry
Use fwts_safe_memcpy rather than a fwts_safe_memread and then an
un-safe copy as a micro optimisation.  Also throw an error and
return a NULL address if the memcpy failed because the memory is
unreadable.  Fixes a bug on aarch64 systems where this entry is
non-readable and causes a BUS error.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>