acpica: fwts_acpica: fix linker issues on Fedora (LP: #1712572)
When building for ppcC64el on Fedora we get linker issues with
undefined symbols. It seems that while this is not an issue on
Debian systems, the linker seems to require these sysbols to be
defined even if we don't actually use them. Make these functions
void no-ops that throw warnings if they are used.  This has been
tested on Fedora 25 and also Ubuntu and is able to build and
pass the regression tests.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>