lib: fwts_framework: re-order and clean up fwts_framework fields
The fields in the fwts_framework structure have historically been added
overtime at the end of the strucutre one by one and the order is now a
bit confusing. Re-order the fields to make it a bit easier to read. Also
make the magic field 64 bits to ensure we have more magic bits to check
on for structure corruption checking.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>