uefi: uefirttime: fix the time zone to exact value
UEFI spec defined the time zone value for -1440 to 1440 and 2047(unspecified
time zone). Using timezone 1 is allowed by UEFI spec, however value 1 is not
exact value on earth. Some firmware will check the exact timezone value on
earth, so the test will fail. Modify the timezone value test with 0 and 2047
for testing.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>