uefi: uefidump: add some guarding on allocation size
Static analysis with CoverityScan is warning about a possible allocation
of an untrusted size calculated from u->dev_path.length.  Add some extra
checking on this.  This does not remove the warning but I'm sure it's as
good as we can to check untrusted data warnings from CoverityScan.
Also remove a whitespace.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>