acpi: bert: remove confusing warning
commit a82d373f65ae ("acpi: bert: do not fail if BERT error region
raw_data_length is zero") added a warning which triggers when a BERT table
is present, but contains no raw data (ie a boot error did not occur).  This
is a perfectly valid scenario, and the spec defines bits which indicate if
valid data is present to be parsed.

The warning is confusing because it implies that having a BERT table which
contains no data is wrong, ie violates the spec, but unlikely to cause a real
issue.  This is incorrect as the spec specifically allows a BERT with no
data.  Additionally, the warning is coupled with advice text which states
that the BERT may not have raw data, ie doing the right thing, which adds
further confusion.

The warning just causes unnecessary churn as questions arise about if the
implementation is following the spec ("yes it is, just ignore the warning"
is hardly satisfactory), and there is no real way to determine if there is
raw data which is not being properly reported to the HLOS, therefore remove
the warning.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>