acpi: fadt: Fix FACS checking
In case of HW reduced platforms, FIRMWARE_CTRL and X_FIRMWARE_CTRL are
allowed to be 0 in the FADT as a FACS may not be provided.  The FACS
validation which checks these fields does register a success for HW reduced
platforms when the fields are both 0, or registers a failure for non-HW
reduced platforms as FACS is required in that case.  However, the FACS
validation continues with other checks, which can override the previous
success with a failure, causing an invalid test result and confusing output
messages.

We can do better by reorganizing the checks to follow in logical succession,
only applying pass/fail results appropiate to the specific scenario (both
fields zero, both fields non-zero, one field zero with the other non-zero).

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>